Update to use TensorRT 10 and support newer graphics cards (#32)
## Motivation
We want to support graphics cards of the 50XX series.
These are not supported by the current TensorRT / CUDA combination, so
we need to update to TensorRT 10.X.
Along with this we want to simplify package versioning and installation
using uv and add some linter configurations.

## Implementation
For TRT10 compatibility I copied the relavant files from the trt-10
branch of tensorrtx.
I also removed all of the non-relevant build targets to speed up the
docker builds.
We now use a pyproject and uv.lock for package management and read the
nodelib version from the pyproject.toml docker.sh.
For simplification, I removed the opencv code and replaced it with a PIL
equivalent (both use Bilinear interpolation for downsizing).

## Changes
- add tensorrtx files to repo
- update yolov5.py to use new engine syntax
- use nvidia tensorrt image as cloud base
- use l4t 36.x based image as jetson base (needs JetPack 6)
- add pyproject.toml for package definitions and linter configuration
- create uv.lock to freeze dependencies
- drop opencv dependency and replace the two uses with PIL and numpy

---------

Co-authored-by: Paula Kammler <paula@zauberzeug.com>
Co-authored-by: Dr. Dennis Wittich <dennis@zauberzeug.com>